+++ compaction_garbage_collector.hh
@@ -22,7 +22,7 @@
 #pragma once
 
 #include "schema.hh"
-#include "types/collection.hh"
+#include "collection_mutation.hh"
 
 class atomic_cell;
 class row_marker;
@@ -31,6 +31,6 @@ class compaction_garbage_collector {
 public:
     virtual ~compaction_garbage_collector() = default;
     virtual void collect(column_id id, atomic_cell) = 0;
-    virtual void collect(column_id id, collection_type_impl::mutation) = 0;
+    virtual void collect(column_id id, collection_mutation_description) = 0;
     virtual void collect(row_marker) = 0;
 };
