+++ cql3/constants.cc
@@ -166,7 +166,7 @@ constants::literal::prepare(database& db, const sstring& keyspace, ::shared_ptr<
 
 void constants::deleter::execute(mutation& m, const clustering_key_prefix& prefix, const update_parameters& params) {
     if (column.type->is_multi_cell()) {
-        collection_type_impl::mutation coll_m;
+        collection_mutation_description coll_m;
         coll_m.tomb = params.make_tombstone();
         auto ctype = static_pointer_cast<const collection_type_impl>(column.type);
         m.set_cell(prefix, column, atomic_cell_or_collection::from_collection_mutation(ctype->serialize_mutation_form(coll_m)));
