+++ cql3/maps.cc
@@ -285,8 +285,8 @@ maps::setter::execute(mutation& m, const clustering_key_prefix& row_key, const u
         return;
     }
     if (column.type->is_multi_cell()) {
-        // delete + put
-        collection_type_impl::mutation mut;
+        // Delete all cells first, then put new ones
+        collection_mutation_description mut;
         mut.tomb = params.make_tombstone_just_before();
         auto ctype = static_pointer_cast<const map_type_impl>(column.type);
         auto col_mut = ctype->serialize_mutation_form(std::move(mut));
@@ -312,7 +312,7 @@ maps::setter_by_key::execute(mutation& m, const clustering_key_prefix& prefix, c
     }
     auto ctype = static_pointer_cast<const map_type_impl>(column.type);
     auto avalue = value ? params.make_cell(*ctype->get_values_type(), *value, atomic_cell::collection_member::yes) : params.make_dead_cell();
-    map_type_impl::mutation update;
+    collection_mutation_description update;
     update.cells.emplace_back(std::move(to_bytes(*key)), std::move(avalue));
     // should have been verified as map earlier?
     auto col_mut = ctype->serialize_mutation_form(std::move(update));
@@ -333,11 +333,11 @@ maps::do_put(mutation& m, const clustering_key_prefix& prefix, const update_para
         shared_ptr<term> value, const column_definition& column) {
     auto map_value = dynamic_pointer_cast<maps::value>(value);
     if (column.type->is_multi_cell()) {
-        collection_type_impl::mutation mut;
-
         if (!value) {
             return;
         }
+
+        collection_mutation_description mut;
 
         auto ctype = static_pointer_cast<const map_type_impl>(column.type);
         for (auto&& e : map_value->map) {
@@ -367,7 +367,7 @@ maps::discarder_by_key::execute(mutation& m, const clustering_key_prefix& prefix
     if (key == constants::UNSET_VALUE) {
         throw exceptions::invalid_request_exception("Invalid unset map key");
     }
-    collection_type_impl::mutation mut;
+    collection_mutation_description mut;
     mut.cells.emplace_back(*key->get(params._options), params.make_dead_cell());
     auto mtype = static_cast<const map_type_impl*>(column.type.get());
     m.set_cell(prefix, column, mtype->serialize_mutation_form(mut));
