+++ mutation_compactor.hh
@@ -97,7 +97,7 @@ public:
     virtual void collect(column_id id, atomic_cell cell) override {
         _row.apply(_schema.column_at(_kind, id), std::move(cell));
     }
-    virtual void collect(column_id id, collection_type_impl::mutation mut) override {
+    virtual void collect(column_id id, collection_mutation_description mut) override {
         if (mut.tomb || !mut.cells.empty()) {
             const auto& cdef = _schema.column_at(_kind, id);
             auto& ctype = *static_pointer_cast<const collection_type_impl>(cdef.type);
