+++ tests/cql_query_test.cc
@@ -2803,7 +2803,7 @@ SEASTAR_TEST_CASE(test_frozen_collections) {
         }).then([&e] {
             return e.execute_cql("SELECT * FROM tfc;");
         }).then([&e, frozen_map_of_set_and_list] (shared_ptr<cql_transport::messages::result_message> msg) {
-            map_type_impl::mutation_view empty_mv{};
+            collection_mutation_view_description empty_mv{};
             assert_that(msg).is_rows().with_rows({
                 { int32_type->decompose(0), int32_type->decompose(0), frozen_map_of_set_and_list->to_value(empty_mv, cql_serialization_format::internal()), int32_type->decompose(0) },
             });
